Added alert for timeouts
Added alert for timeouts

This is the place where the 500 errors are thrown during the import. From several tests i’ve found out that the items might or not get all through.. So it’s not safe to assume that the import failed completely, because there will still be items loaded until the point where it timed out, and even then, the timeout could happen right at the end..

But if the timeout happens with the last media item, the pages/posts don’t get imported anyway.

So the alert makes the user aware that there COULD (and in 99% cases it’s true) be an error.